{"schema":"decodex/commit/1","summary":"Modularize service acceptance test helpers","authority":"manual"}
