+++ packages/elf-service/tests/acceptance/knowledge_pages/helpers/source_inserts/note.rs
@@ -1,0 +1,29 @@
+use uuid::Uuid;
+
+use crate::acceptance::knowledge_pages::helpers::{AGENT_ID, PROJECT_ID, TENANT_ID};
+use elf_service::{AddNoteInput, AddNoteRequest, ElfService};
+
+pub(super) async fn insert_source_note(service: &ElfService, key: &str, text: &str) -> Uuid {
+	let response = service
+		.add_note(AddNoteRequest {
+			tenant_id: TENANT_ID.to_string(),
+			project_id: PROJECT_ID.to_string(),
+			agent_id: AGENT_ID.to_string(),
+			scope: "agent_private".to_string(),
+			notes: vec![AddNoteInput {
+				r#type: "fact".to_string(),
+				key: Some(key.to_string()),
+				text: text.to_string(),
+				structured: None,
+				importance: 0.7,
+				confidence: 0.9,
+				ttl_days: None,
+				source_ref: serde_json::json!({ "schema": "acceptance/v1", "key": key }),
+				write_policy: None,
+			}],
+		})
+		.await
+		.expect("add_note should persist source note");
+
+	response.results[0].note_id.expect("source note id should be present")
+}
